ref: move test case to /lite/test
