+++ lite/tests/benchmark/src/elementwise_arm_math.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2019 PaddlePaddle Authors. All Rights Reserved.
+// Copyright (c) 2020 PaddlePaddle Authors. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,7 +17,6 @@
 #include <vector>
 
 #include "lite/backends/arm/math/elementwise.h"
-#include "lite/core/op_registry.h"
 
 static void fast_bcast_args(benchmark::internal::Benchmark* b) {
   b->ArgNames({"batch", "channel", "num"});
